[FIX] custom_background: properly inherit _run_wkhtmltopdf (#28)
* [FIX] custom_background: properly inherit _run_wkhtmltopdf

As it is just post-processing of the PDF generated by Odoo Core, we convert `pdf_content` to
a file and write it again to the temp directory and proceed as done before without needlessly overriding the whole method.

* [FIX] custom_background: only apply specific print media settings to PDF

As we want to apply only print media style to wkhtmltopdf we have to inject the command
and adapt the scss accordingly. The background should not be transparent in the screen view.